+++ docs/CHANGELOG.rst
@@ -13,6 +13,7 @@ Changed
 -------
 - **BACKWARD INCOMPATIBLE:** Require Resolwe 21.x
 - Use human samples for ``alignment-star`` testing
+- Bump memory requirements in ``rrbs-metrics`` process
 
 Fixed
 -----